% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-6-construct_qc_config.R
\name{add_custom_config_per_sample}
\alias{add_custom_config_per_sample}
\title{Customize configuration for each sample in an experiment}
\usage{
add_custom_config_per_sample(
  customize_template_config,
  config_template,
  scdata_list,
  unfiltered_samples = NA
)
}
\arguments{
\item{customize_template_config}{function - step customization function}

\item{config_template}{list - template of step configuration parameters}

\item{scdata_list}{list - with Seurat objects}

\item{unfiltered_samples}{character vector}
}
\value{
list of customized QC parameters for each sample
}
\description{
Takes the step config template and the required data and calls the corresponding
step customization function, which takes care of changing the QC parameters to
sensible values for each sample in the running experiment.
}
\details{
Values that apply to all experiments are defined in the \code{data-raw/processing_config_template.R}
file.
}
