% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{add_metadata}
\alias{add_metadata}
\title{Add the metadata present in scdata_list into the merged SeuratObject}
\usage{
add_metadata(scdata, scdata_list)
}
\arguments{
\item{scdata}{SeuratObject}

\item{scdata_list}{list of sample SeuratObjects}
}
\value{
SeuratObject with added metadata
}
\description{
This function adds metadata, some of which is present in the sample SeuratObjects
(the experimentID, the gene annotations), and some that is not (the color pool)
to the merged SeuratObject.
}
