% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-X-subset_experiment.R
\name{add_new_sample_ids}
\alias{add_new_sample_ids}
\title{Add new sample ids to the subset Seurat Object}
\usage{
add_new_sample_ids(scdata, sample_id_map)
}
\arguments{
\item{scdata}{Seurat Object}

\item{sample_id_map}{data.table of parent/subset sample id map}
}
\value{
SeuratObject with new sample ids
}
\description{
Add new sample ids to the subset Seurat Object
}
