% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-1-subset_seurat.R
\name{add_subset_metadata}
\alias{add_subset_metadata}
\title{add experiment level metadata to subset seurat object}
\usage{
add_subset_metadata(input, subset_scdata, sample_id_map)
}
\arguments{
\item{input}{list of input params, containing the experimentId}

\item{sample_id_map}{list with mapping between sample_ids from
parent and subset experiments}

\item{scdata}{seurat object}
}
\value{
scdata with additional metadata
}
\description{
add experiment level metadata to subset seurat object
}
