% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-functions.R
\name{build_activity_arn}
\alias{build_activity_arn}
\title{Build activity ARN}
\usage{
build_activity_arn(aws_region, aws_account_id, activity_id)
}
\arguments{
\item{aws_region}{character}

\item{aws_account_id}{character}

\item{activity_id}{character}
}
\value{
string of activity ARN
}
\description{
Build activity ARN
}
