% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{build_cc_gene_list}
\alias{build_cc_gene_list}
\title{Make list of cell cycle genes}
\usage{
build_cc_gene_list(all_genes)
}
\arguments{
\item{all_genes}{character vector, gene_annotations$input}
}
\value{
integer vector of cell cycle gene indices
}
\description{
Uses cell cycle gene list from Tirosh et. al. 2016, bundled with Seurat.
}
\details{
For now it's only useful for Homo sapiens. But could be easily extended to mice
converting the human gene names to their orthologs in mice, as suggested by
satijalab in \href{https://github.com/satijalab/seurat/issues/2493#issuecomment-575702274}{this github issue}
}
