% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-7-upload_to_aws.R
\name{build_metadata_cellsets}
\alias{build_metadata_cellsets}
\title{Create cellsets from user-supplied metadata}
\usage{
build_metadata_cellsets(input, scdata_list, color_pool)
}
\arguments{
\item{input}{list}

\item{scdata_list}{list of Seurat objects}

\item{color_pool}{list of colors to use}
}
\value{
list of cellsets
}
\description{
This function creates the cellsets for the user-supplied metadata (in data
management module). It also takes care of assigning a color from the color pool
to each cellset.
}
