% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{build_mitochondrial_gene_list}
\alias{build_mitochondrial_gene_list}
\title{Make list of mitochondrial genes}
\usage{
build_mitochondrial_gene_list(all_genes)
}
\arguments{
\item{all_genes}{character vector, gene_annotations$input}
}
\value{
integer vector of mitochondrial gene indices
}
\description{
Matches mitochondrial genes using the same regular expression as used in QC's
mitochondrial filter.
}
