% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{build_ribosomal_gene_list}
\alias{build_ribosomal_gene_list}
\title{Make list of ribosomal genes}
\usage{
build_ribosomal_gene_list(all_genes)
}
\arguments{
\item{all_genes}{character vector, gene_annotations$input}
}
\value{
integer vector of ribosomal gene indices
}
\description{
Matches ribosomal gene symbols using a regular expression that covers most
cases across several commonly used species. It also matches 3 extra ribosomal
genes for human and mouse, which are not covered by the regex.
}
