% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-7-upload_to_aws.R
\name{build_scratchpad_cellsets}
\alias{build_scratchpad_cellsets}
\title{Create cell set using scratchpad information from parent cell set}
\usage{
build_scratchpad_cellsets(color_pool, child_cellsets)
}
\arguments{
\item{color_pool}{list of colors to use}

\item{child_cellsets}{cell set resulting from parent cell set filtering}
}
\value{
cell set filled with scratchpad information from the parent cell set
}
\description{
This function create a cell set with scratchpad information including only
the cell ids in the subset experiment. If all cell ids from a scratchpad cluster
of the parent experiments are filtered out in the subset experiment,
then that scratchpad cluster is not included in the subset cell set.
}
