% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-functions.R
\name{call_copy}
\alias{call_copy}
\title{Call copy}
\usage{
call_copy(task_name, input, pipeline_config)
}
\arguments{
\item{task_name}{character name of the step}

\item{input}{list containing
- parentExperimentId
- childExperimentId
- sample IDs, and names}

\item{pipeline_config}{list as defined by load_config}
}
\value{
character message id
}
\description{
Runs step `task_name` of the copy pipeline, sends output message to the API
}
