% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-functions.R
\name{call_gem2s}
\alias{call_gem2s}
\title{Call gem2s}
\usage{
call_gem2s(task_name, input, pipeline_config)
}
\arguments{
\item{task_name}{character name of the step}

\item{input}{list containing
\itemize{
\item experimentID
\item sample IDs, names and S3 paths
}}

\item{pipeline_config}{list as defined by load_config}
}
\value{
character message id
}
\description{
Runs step \code{task_name} of the GEM2S pipeline, sends output message to the API
}
