% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-functions.R
\name{call_qc}
\alias{call_qc}
\title{Call QC pipeline}
\usage{
call_qc(task_name, input, pipeline_config)
}
\arguments{
\item{task_name}{character name of the step}

\item{input}{list containing:
- step parameters for all samples
- current sample UUID
- uploadCountMatrix (whether or not to upload matrix after step)}

\item{pipeline_config}{list as defined by load_config}
}
\value{
character message id
}
\description{
Runs step `task_name` of the data processing pipeline, sends plot data to s3
and the output message to the API
}
