% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_cellranger.R
\name{call_read10x}
\alias{call_read10x}
\title{Calls Read10X}
\usage{
call_read10x(config)
}
\arguments{
\item{config}{experiment settings.}
}
\value{
list with an element per sample with dgCMatrix with genes as rows and cells as column
}
\description{
Cellranger outputs from V2 and V3 kits were renamed to look like V3 (features.tsv.gz).
}
