% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_cellranger_files.R
\name{call_read_rhapsody}
\alias{call_read_rhapsody}
\title{Calls BD rhapsody data parsing functions}
\usage{
call_read_rhapsody(config, input_dir)
}
\value{
list containing \itemize{
\item{"counts_list"}{named list of dgCMatrix per sample}
\item{"annot"}{data.frame with gene symbols}
}
\description{
Currently we only have support for sparse expression matrices.
}
