% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-functions.R
\name{call_subset}
\alias{call_subset}
\title{Call subset seurat}
\usage{
call_subset(task_name, input, pipeline_config)
}
\arguments{
\item{task_name}{character name of the step}

\item{input}{list containing
\itemize{
\item parentExperimentId
\item childExperimentId
\item sample IDs, and names
}}

\item{pipeline_config}{list as defined by load_config}
}
\value{
character message id
}
\description{
Runs step \code{task_name} of the subset seurat pipeline, sends output message to the API
}
