% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-4-score_doublets.R
\name{compute_doublet_scores}
\alias{compute_doublet_scores}
\title{Compute doublets scores per sample.}
\usage{
compute_doublet_scores(counts)
}
\arguments{
\item{scdata}{Raw sparse matrix with the counts for one sample.}
}
\value{
data.frame with doublet scores and assigned classes
}
\description{
Compute doublets scores per sample.
}
