% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-3-run_emptydrops.R
\name{compute_sample_edrops}
\alias{compute_sample_edrops}
\title{Calculate empty drops scores for sample}
\usage{
compute_sample_edrops(sample_counts)
}
\arguments{
\item{sample_counts}{dgCMatrix with counts for one sample.}
}
\value{
data.frame with edrops scores
}
\description{
Calculate empty drops scores for sample
}
