% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-5-create_seurat.R
\name{construct_metadata}
\alias{construct_metadata}
\title{Construct metadata for each SeuratObject}
\usage{
construct_metadata(counts, sample, config)
}
\arguments{
\item{counts}{count matrix}

\item{sample}{character sample ID}

\item{config}{list containing experiment config}
}
\value{
data.frame of sample metadata
}
\description{
This function creates a \code{data.frame} with the barcodes, sampleIDs and user supplied
metadata that corresponds to each one.
}
