% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-1-subset_seurat.R
\name{create_sample_id_map}
\alias{create_sample_id_map}
\title{generate a sample id mapping for remaining samples after subset}
\usage{
create_sample_id_map(parent_sample_id)
}
\arguments{
\item{parent_sample_id}{character vector of unique parent sample ids}
}
\value{
data.table with sample id map
}
\description{
New sample ids must be created, but the number of samples depends on which
cells have been subset by the user. Sample Ids that belong to the parent
experiment are also kept, which is useful for the addition of the new subclusters
to the parent experiment.
}
