% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{create_scdata}
\alias{create_scdata}
\title{Create the merged Seurat object}
\usage{
create_scdata(scdata_list, cells_id)
}
\arguments{
\item{scdata_list}{list of SeuratObjects}
}
\value{
SeuratObject
}
\description{
This function takes care of merging the sample seurat objects, shuffling
and adding the metadata to the complete Seurat object. It does so by calling
the corresponding functions.
}
