% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-5-create_seurat.R
\name{create_seurat}
\alias{create_seurat}
\title{Create a SeuratObject per sample}
\usage{
create_seurat(input, pipeline_config, prev_out)
}
\arguments{
\item{input}{The input object from the request}

\item{pipeline_config}{result of \code{load_config}}

\item{prev_out}{'output' slot from call to \code{score_doublets}}
}
\value{
\code{prev_out} with added slot 'scdata_list' containing \code{SeuratObject}'s for each sample.
}
\description{
Create a SeuratObject per sample
}
