% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-1-download_cellranger.R
\name{download_cellranger}
\alias{download_cellranger}
\title{Download cellranger files from S3}
\usage{
download_cellranger(input, pipeline_config, prev_out = list())
}
\arguments{
\item{input}{The input object from the request}

\item{pipeline_config}{result of \code{load_config}}

\item{prev_out}{\code{list()} that is added to in each with gem2s task.}
}
\value{
list where 'output' slot has config used by \code{load_cellranger}
}
\description{
Download cellranger files from S3
}
