% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-7-embed_and_cluster.R
\name{download_cl_metadata_file}
\alias{download_cl_metadata_file}
\title{Download and load cell-level metadata tsv file}
\usage{
download_cl_metadata_file(config)
}
\arguments{
\item{config}{list with AWS configuration and metadataS3Path}
}
\value{
data.table of cell-level metadata
}
\description{
Download and load cell-level metadata tsv file
}
