% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-1-download_user_files.R
\name{download_user_files}
\alias{download_user_files}
\title{Download user files from S3}
\usage{
download_user_files(input, pipeline_config, prev_out = list())
}
\arguments{
\item{input}{The input object from the request}

\item{pipeline_config}{result of \code{load_config}}

\item{prev_out}{\code{list()} that is added to in each with gem2s task.}
}
\value{
list where 'output' slot has config used by \code{load_user_files}
}
\description{
Download user files from S3
}
