% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-7-embed_and_cluster.R
\name{embed_and_cluster}
\alias{embed_and_cluster}
\title{run clustering}
\usage{
embed_and_cluster(
  scdata,
  config,
  sample_id,
  cells_id,
  task_name = "configureEmbedding",
  ignore_ssl_cert = FALSE
)
}
\arguments{
\item{scdata}{seurat object}

\item{config}{list with clustering parameters}

\item{sample_id}{character}

\item{cells_id}{list of cell ids that passed all filters so far}

\item{task_name}{character}
}
\value{
list
}
\description{
run clustering
}
