% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-helpers.R
\name{ensure_is_list_in_json}
\alias{ensure_is_list_in_json}
\title{Ensure is list in json}
\usage{
ensure_is_list_in_json(vector)
}
\arguments{
\item{vector}{}
}
\description{
When sending responses as json, Vectors of length 0 or 1 are converted to
null and scalar (respectively) Using as.list fixes this, however, long R
lists take a VERY long time to be converted to JSON.
This function deals with the problematic cases, leaving vector as a vector
when it isnt a problem.
}
