% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-7-upload_to_aws.R
\name{extract_subset_user_metadata}
\alias{extract_subset_user_metadata}
\title{create list of user supplied metadata tracks and values}
\usage{
extract_subset_user_metadata(subset_cellsets)
}
\arguments{
\item{subset_cellsets}{data.table of cellsets}
}
\value{
list of metadata tracks and values
}
\description{
create list of user supplied metadata tracks and values
}
