% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-5-filter_doublets.R
\name{filter_doublets}
\alias{filter_doublets}
\title{STEP 5. Doublet score filter}
\usage{
filter_doublets(
  scdata_list,
  config,
  sample_id,
  cells_id,
  task_name = "doubletScores",
  num_cells_to_downsample = 6000
)
}
\arguments{
\item{config}{list containing the following information
- enable: true/false. Refering to apply or not the filter.
- auto: true/false. 'True' indicates that the filter setting need to be changed depending on some sensible value (it requires
to call generate_default_values_doubletScores)
- filterSettings: slot with thresholds
- probabilityThreshold: Float. cut-off for the maximun probability scores that could have a cell. The doubletScores scores have been computed
through "scrublet" \link{1}
- binStep: Float. Bin size for the histogram}
}
\value{
a list with the filtered seurat object by doublet score, the config and the plot values
}
\description{
Filters seurat object based on doubletScores scores
}
\details{
This is a simplest filter that looks at a threshold value for the doublet scores.
To separate cells with low droplet score from the ones that have a high droplet score content what makes us think that the are mistakenly considered as a single cell but they are actully two or more.
This can be a useful first guess. The settings for such a filter can also contain a simple "probabilityThreshold" setting.
}
