% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-1-subset_seurat.R
\name{filter_low_cell_samples}
\alias{filter_low_cell_samples}
\title{Remove samples with low cell numbers}
\usage{
filter_low_cell_samples(scdata, min_cells = MIN_CELLS_IN_SAMPLE)
}
\arguments{
\item{scdata}{Seurat Object}

\item{min_cells}{integer of minimum number of cells for a sample to be kept}
}
\value{
Seurat object with samples containing at least min_cells
}
\description{
When subsetting, there's an increased likelihood of having samples with low
cell numbers. This breaks the pipeline in different places, especially during
integration. This function removes samples with cell numbers below an empirically
determined threshold
}
