% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-2-filter_low_cellsize.R
\name{filter_low_cellsize}
\alias{filter_low_cellsize}
\title{STEP 2. Cell size distribution filter}
\usage{
filter_low_cellsize(
  scdata_list,
  config,
  sample_id,
  cells_id,
  task_name = "cellSizeDistribution",
  num_cells_to_downsample = 6000
)
}
\arguments{
\item{config}{list containing the following information
- enable: true/false. Refering to apply or not the filter.
- auto: true/false. 'True' indicates that the filter setting need to be changed depending on some sensible value (it requires
to call generate_default_values_cellSizeDistribution)
- filterSettings: slot with thresholds
- minCellSize: Integer. Threshold that contain the minimun number of UMIs per cell
- binStep: Integer. Bin size for the histogram}
}
\value{
a list with the filtered seurat object by cell size ditribution, the config and the plot values
}
\description{
cell_size_distribution_filter function
}
\details{
This is a simplest filter that looks at the shape of the cell size (# of UMIs
per cell) distribution and looks for some local minima, minimum of second derivative, etc.
To separate the large cell barcodes that correspond to real cells from the
tail containing 'empty droplets'. This can be a useful first guess. The settings
for such a filter can also contain a simple "min cell size" setting.
}
