% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-7-upload_to_aws.R
\name{filter_parent_cellsets}
\alias{filter_parent_cellsets}
\title{Filter parent cellsets}
\usage{
filter_parent_cellsets(parent_cellsets, cell_ids_to_keep)
}
\arguments{
\item{parent_cellsets}{cell set object from the parent experiment}

\item{cell_ids_to_keep}{vector of cell ids to keep}
}
\value{
filtered cell set object
}
\description{
This function filterz the parent cellset removing all the information about
louvain/leiden clusters because they will be recalculated in the subset experiment.
It removes also all the cell ids from the parent experiment that are not present
in the subset experiment.
}
