% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{filter_unnamed_features}
\alias{filter_unnamed_features}
\title{Try to fix genes without name or remove them}
\usage{
filter_unnamed_features(counts, annotations, sample)
}
\arguments{
\item{counts}{count matrix}

\item{annotations}{list of annotations data.frame, feature types and gene_column}

\item{sample}{character specifying current sample}
}
\value{
list of counts and annotations
}
\description{
This function checks if genes whose annotations in the count matrix (\code{rownames(counts)})
are empty, can be annotated with a different column from the features file,
read into the \code{annotations[["annot"]]} table. If it can, they are replaced everywhere
(both in the count matrix and in all columns of the annotations table). In case
the other columns do not provide a better annotation, they are removed from
the count matrix and annotations table.
}
