% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-7-embed_and_cluster.R
\name{find_clm_columns}
\alias{find_clm_columns}
\title{check if variable is acceptable as cell-level metadata}
\usage{
find_clm_columns(check_vals)
}
\arguments{
\item{check_vals}{vector of arbitrary type}
}
\value{
boolean specifying if variable is desirable or not
}
\description{
Tests values using several heurisitics to determine if a variable has a reasonable
number of values to be made into cellsets. Cellsets are categorical, therefore
continuous or high-cardinality variables are filtered out.
}
