% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-7-embed_and_cluster.R
\name{format_cl_metadata_cellsets}
\alias{format_cl_metadata_cellsets}
\title{Formats cell-level metadata cellsets}
\usage{
format_cl_metadata_cellsets(cl_metadata, var_types, color_pool)
}
\arguments{
\item{cl_metadata}{data.table}

\item{var_types}{list of variable types}

\item{color_pool}{list}
}
\value{
list of cell-level metadata cell classes
}
\description{
Formats cell-level metadata cellsets
}
