% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-1-filter_emptydrops.R
\name{generate_default_values_classifier}
\alias{generate_default_values_classifier}
\title{Generate default values for classifier filter}
\usage{
generate_default_values_classifier(scdata, config)
}
\arguments{
\item{config}{list containing the following information
- enable: true/false. Refering to apply or not the filter.
- auto: true/false. 'True' indicates that the filter setting need to be changed depending on some sensible value (it requires
to call generate_default_values_classifier)
- filterSettings: slot with thresholds
- minProbabiliy:
- filterThreshold:}
}
\value{
a list with the filtered seurat object by probabilities classifier, the config and the plot values
}
\description{
Filters seurat object based on classifier filter using emptyDrops
}
