% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{generate_elbow_plot_data}
\alias{generate_elbow_plot_data}
\title{generate elbow plot data}
\usage{
generate_elbow_plot_data(scdata_integrated, config, task_name, var_explained)
}
\arguments{
\item{scdata_integrated}{integrated seurat object}

\item{config}{list}

\item{task_name}{character}

\item{var_explained}{numeric}
}
\value{
list of plot data
}
\description{
Reshapes table to an UI compatible format for elbow/scree plot.
}
