% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-helpers.R
\name{generate_first_step_ids}
\alias{generate_first_step_ids}
\title{Generate the list of cell ids present in the experiment}
\usage{
generate_first_step_ids(scdata_list)
}
\arguments{
\item{scdata_list}{list of Seurat Objects}
}
\value{
list of cell ids present in the Seurat object
}
\description{
Generate the list of cell ids present in the experiment
}
