% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-1-subset_seurat.R
\name{generate_subset_config}
\alias{generate_subset_config}
\title{Translate processing config}
\usage{
generate_subset_config(parent_processing_config, sample_id_map)
}
\arguments{
\item{sample_id_map}{Maps parent sample ids to the subset sample ids
(only the parent samples that have cells after the subset are included)}

\item{parentProcessingConfig}{The processingConfig of the parent experiment}
}
\value{
Processing config for the subset experiment
}
\description{
Auxiliary function for subset_seurat
}
