% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-2-filter_low_cellsize.R
\name{get_bcranks_plot_data}
\alias{get_bcranks_plot_data}
\title{Get Barcode Ranks Plot Data}
\usage{
get_bcranks_plot_data(sample_subset, nmax = 6000, is.cellsize = TRUE)
}
\arguments{
\item{sample_subset}{\code{SeuratObject}}

\item{nmax}{maximum number of value for histogram plot. Only applies if \code{is.cellsize}
is \code{TRUE}.}

\item{is.cellsize}{is this for cellsize filter? Default (\code{TRUE}),
returns histogram plot data and knee plot data without fdr and ndrops values.}
}
\value{
Named list of plot data. Possible items are \code{'hist'} and \code{'knee'},
dependending on \code{is.cellsize}.
}
\description{
Get Barcode Ranks Plot Data
}
\keyword{internal}
