% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-7-embed_and_cluster.R
\name{get_cell_id_barcode_map}
\alias{get_cell_id_barcode_map}
\title{extract cell_id - barcode map}
\usage{
get_cell_id_barcode_map(scdata)
}
\arguments{
\item{scdata}{SeuratObject}
}
\value{
data.table containing cells_id, barcode, samples
}
\description{
Helper to get required variables from the SeuratObject
}
