% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-7-upload_to_aws.R
\name{get_cell_sets}
\alias{get_cell_sets}
\title{Create initial cell sets object}
\usage{
get_cell_sets(scdata, input)
}
\arguments{
\item{scdata}{list of Seurat objects}

\item{input}{The input object from the request}
}
\value{
cell set object
}
\description{
Create initial cell sets object
}
