% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{get_feature_types}
\alias{get_feature_types}
\title{Determine the type of features in the annot data frame}
\usage{
get_feature_types(annot)
}
\arguments{
\item{annot}{data.frame read from features file}
}
\value{
-1 is SYMBOL/IDS
0 is SYMBOL/ SYMBOL
1 is IDS/SYMBOL
2 is IDS/IDS
}
\description{
Classifies the columns into either ensemblIds or symbols, and extracts a
number that represents the combination.
}
