% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-1-download_user_files.R
\name{get_gem2s_file}
\alias{get_gem2s_file}
\title{Download user files from S3}
\usage{
get_gem2s_file(input, originals_bucket, input_dir, s3)
}
\arguments{
\item{input}{The input params received from the api}

\item{originals_bucket}{The bucket where the sample files are}

\item{input_dir}{A string, where the s3 object is going to be stored}

\item{s3}{S3 aws client}

\item{aws_config}{The input object from the request}
}
\description{
Download user files from S3
}
