% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-7-upload_to_aws.R
\name{get_subset_cell_sets}
\alias{get_subset_cell_sets}
\title{Create cell set object for subset experiment}
\usage{
get_subset_cell_sets(
  scdata_list,
  input,
  prev_out,
  disable_qc_filters,
  sample_id_map
)
}
\arguments{
\item{scdata_list}{list of Seurat objects}

\item{input}{The input object from the request}

\item{prev_out}{list with results appended in each gem2s task}

\item{disable_qc_filters}{bool indicating if the data derives from the
subsetting of another experiment}
}
\value{
cell set object
}
\description{
Create cell set object for subset experiment
}
