% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-helpers.R
\name{getClusters}
\alias{getClusters}
\title{Get Clusters}
\usage{
getClusters(clustering_method, resolution, data)
}
\arguments{
\item{clustering_method}{character}

\item{resolution}{numeric}

\item{data}{Seurat Object}
}
\value{
Seurat object with clustering assignment
}
\description{
Get Clusters
}
