% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{ids_to_sym}
\alias{ids_to_sym}
\title{Convert IDS to symbols using key data.frame}
\usage{
ids_to_sym(sample_annot, annot_with_ids)
}
\arguments{
\item{sample_annot}{data.frame of annotations}

\item{annot_with_ids}{data.frame of annotations with IDs. Key data.frame}
}
\value{
data.frame of annotations
}
\description{
This function tries to convert IDs to symbols using the key data.frame created
with \code{make_annot_with_ids}. In case there's no symbol available, it keeps the
existing ID
}
