% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{integrate_from_sketch}
\alias{integrate_from_sketch}
\title{Integrate using sketch data}
\usage{
integrate_from_sketch(scdata, scdata_sketch, integration_function, config)
}
\arguments{
\item{scdata}{Seurat object - complete data set}

\item{scdata_sketch}{Seurat object - sketch}

\item{integration_function}{function}

\item{config}{list - integration parameters}

\item{method}{character - integration methods}
}
\value{
Integrated Seurat object
}
\description{
This function takes the sketched data, integrates it and then transfers the
integration to the complete data set.
}
