% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{integrate_scdata}
\alias{integrate_scdata}
\title{STEP 6. Data Integration}
\usage{
integrate_scdata(
  scdata_list,
  config,
  sample_id,
  cells_id,
  task_name = "dataIntegration",
  use_geosketch = FALSE,
  perc_num_cells = 5
)
}
\arguments{
\item{config}{list containing the following information
- dataIntegration
- method: String. Method to be used. "harmony by default.
- methodSettings: List with the method as key and:
- numGenes: Numeric. Number of gene to be used.
- normalisation: String. Normalisation method to be used.
- dimensionalityReduction
- method: String. Method to be used. "rpca" by default.
- numPCs: Numeric. Number of principal components.
- excludeGeneCategories: List. Categories of genes to be excluded.}
}
\value{
a list with the integrated seurat object, the cell ids, the config and the plot values.
}
\description{
Data integration step where batch effect is corrected through data
integration methods. The default method is Harmony.
The data integration step include ordering the samples according to their
matrices size, merging of Seurat objects, normalization and PCA analysis.
Optional: performs geometric sketching of merged Seurat object, integrates
the sketched data, and learn and apply back the integration transformation
to the full data.
}
