% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate-with-seuratv4.R
\name{integrate_using_geosketch}
\alias{integrate_using_geosketch}
\title{Downsample and integrate sketches}
\usage{
integrate_using_geosketch(
  scdata_list,
  cells_id,
  reduction,
  perc_num_cells,
  normalization,
  npcs,
  misc,
  nfeatures,
  use_geosketch
)
}
\arguments{
\item{scdata_list}{list of Seurat objects}

\item{cells_id}{list of cells ids to keep}

\item{reduction}{reduction method}

\item{perc_num_cells}{percentage of cells to keep when using geosketch}

\item{normalization}{normalization method}

\item{npcs}{number of princpal components}

\item{misc}{misc slot from the Seurat object}

\item{nfeatures}{number of features}

\item{use_geosketch}{boolean indicating if geosketch has to be run}
}
\value{
integrated Seurat object
}
\description{
This function uses Geosketch to downsample the dataset, integrates the resulting
sketches using the Seurat v4 workflow, then learns
the integration transformation and applies it to the whole dataset.
}
