% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{list_exclude_genes}
\alias{list_exclude_genes}
\title{Build list of genes to exclude}
\usage{
list_exclude_genes(all_genes, exclude_groups, exclude_custom)
}
\arguments{
\item{all_genes}{character vector, gene_annotations$input}

\item{exclude_groups}{list of groups to exclude}

\item{exclude_custom}{list of custom (user provided) genes to exclude}
}
\value{
integer vector of gene indices to exclude
}
\description{
This function builds the union of gene indices to exclude, joining all groups.
To do so, it calls all the required list_* functions, getting the exclude gene
ids and joins them.
}
