% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{load_cellsets}
\alias{load_cellsets}
\title{Load cellsets object from s3 and flatten it for easier manipulation in R}
\usage{
load_cellsets(s3, pipeline_config, experiment_id)
}
\arguments{
\item{s3}{paws::s3 object}

\item{pipeline_config}{list}

\item{experiment_id}{character}
}
\value{
cellsets list
}
\description{
Load cellsets object from s3 and flatten it for easier manipulation in R
}
