% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj2s-2-load_obj2s_file.R
\name{load_obj2s_file}
\alias{load_obj2s_file}
\title{Loads a Seurat/SingleCellExperiment/AnnData object provided by the user}
\usage{
load_obj2s_file(input, pipeline_config, prev_out, input_dir = "/input")
}
\arguments{
\item{input}{The input params received from the api.}

\item{pipeline_config}{List with S3 bucket name and SNS topics.}

\item{prev_out}{'output' slot from call to \code{download_user_files}}

\item{input_dir}{A string, where the folder containing the downloaded
Seurat object is stored}
}
\description{
This functions is a thin wrapper around \verb{reconstruct_*} functions, which are used to
safely read, check, and reconstruct a Seurat object for downstream use.
}
