% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-1-subset_seurat.R
\name{load_parent_experiment_data}
\alias{load_parent_experiment_data}
\title{load parent experiment data}
\usage{
load_parent_experiment_data(input, pipeline_config)
}
\arguments{
\item{input}{list of input parameters}

\item{pipelne_config}{list of pipeline parameters}
}
\value{
list with scdata and parsed cellsets
}
\description{
Loads the processed rds and cellsets file from the parent experiment from s3.
}
