% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat-2-load_seurat.R
\name{load_seurat}
\alias{load_seurat}
\title{Loads a Seurat object provided by the user}
\usage{
load_seurat(input, pipeline_config, prev_out, input_dir = "/input")
}
\arguments{
\item{input}{The input params received from the api.}

\item{pipeline_config}{List with S3 bucket name and SNS topics.}

\item{prev_out}{'output' slot from call to \code{download_user_files}}

\item{input_dir}{A string, where the folder containing the downloaded
Seurat object is stored}
}
\description{
This functions is a thin wrapper around \code{reconstruct_seurat}, which is used to
safely read, check, and reconstruct the Seurat object for downstream use.
}
