% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{load_user_files}
\alias{load_user_files}
\title{Read user input files}
\usage{
load_user_files(input, pipeline_config, prev_out, input_dir = "/input")
}
\arguments{
\item{input}{The input object from the request}

\item{pipeline_config}{result of \code{load_config}}

\item{prev_out}{list with experiment configuration settings}
}
\value{
list with 'output' slot containing \itemize{
\item{"counts_list"}{named list of dgCMatrix per sample}
\item{"annot"}{data.frame with gene ids and/or symbols}
}
\description{
Checks technology used and dispatches call to correct reading function.
In case of 10x, \code{annot} contains a data.frame from reading the
features.tsv.gz file, while in rhapsody it contains the gene symbols as read
from the count matrix.
}
