% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{log_normalize}
\alias{log_normalize}
\title{Perform log normalization}
\usage{
log_normalize(scdata, normalization_method, integration_method, nfeatures)
}
\arguments{
\item{scdata}{SeuratObject}

\item{normalization_method}{normalization method}

\item{integration_method}{integration method}

\item{nfeatures}{number of features to pass to Seurat::FindVariableFeatures()}
}
\value{
normalized and scaled SeuratObject
}
\description{
#' If the integration method is fastMNN, it will skip ScaleData() because
fastMNN already performs its own scaling.
If the integration method is SeuratV4, the default assay will be set to "integrated",
in this case NormalizeData() will not work (see the \href{https://satijalab.org/seurat/articles/integration_introduction.html}{integration vignette}),
so here it's skipped.
}
