% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{make_annot_with_ids}
\alias{make_annot_with_ids}
\title{Make key annotation data.frame}
\usage{
make_annot_with_ids(annot_list, feature_types_list)
}
\arguments{
\item{annot_list}{list of annotation data.frames}

\item{feature_types_list}{list of feature types}
}
\value{
data.frame of ids and symbols
}
\description{
This function takes the annotation data.frames which contain both symbols and
ids and creates a key data.frame, to be used for conversion of symbols to ids
or viceversa.
}
