% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-7-embed_and_cluster.R
\name{make_cl_metadata_cellclass}
\alias{make_cl_metadata_cellclass}
\title{Transform a variable into a cell-level metadata cell class}
\usage{
make_cl_metadata_cellclass(variable, type, cl_metadata, color_pool)
}
\arguments{
\item{variable}{character - name of variable to transform}

\item{type}{character - type of cell-level metadata variable}

\item{cl_metadata}{data.table}

\item{color_pool}{list}
}
\value{
list - correctly formatted cell-level metadata cellclass
}
\description{
The output is a list that contains one children cellset per value of the
variable to transform.
}
