% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-7-embed_and_cluster.R
\name{make_cl_metadata_cellsets}
\alias{make_cl_metadata_cellsets}
\title{Makes cell-level metadata cellsets}
\usage{
make_cl_metadata_cellsets(scdata, config)
}
\arguments{
\item{scdata}{SeuratObject}

\item{config}{list with AWS config and metadataS3Paths}
}
\value{
list of cell-level metadata cellsets
}
\description{
Main cell-level metadata cellsets function. Downloads, parses, de-duplicates
barcodes, detects variable types and formats cell-level metadata cellsets.
Returning a list ready to add to \code{cellsets.json} file.
}
