% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{merge_scdata_list}
\alias{merge_scdata_list}
\title{Merge the list of sample Seurat objects}
\usage{
merge_scdata_list(scdata_list)
}
\arguments{
\item{scdata_list}{list of SeuratObjects}
}
\value{
SeuratObject
}
\description{
If the list contains more than one seurat object, it merges them all. Else,
returns the only element.
}
