% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-6-prepare_experiment.R
\name{merge_scdatas}
\alias{merge_scdatas}
\title{Merge scdatas: merge rds files in input list}
\usage{
merge_scdatas(scdata_list)
}
\arguments{
\item{scdata_list}{}
}
\value{

}
\description{
Merge scdatas: merge rds files in input list
}
