% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{normalize_annotation_types}
\alias{normalize_annotation_types}
\title{normalize annotation types}
\usage{
normalize_annotation_types(
  annot_list,
  counts_list,
  feature_types_list,
  samples
)
}
\arguments{
\item{annot_list}{list of annotation data.frames}

\item{counts_list}{list of count matrices}

\item{feature_types_list}{list of feature types}

\item{samples}{list of samples}
}
\value{
list with normalized annotations and count matrices
}
\description{
This function makes annotations compatible between samples with different types.
There are three possible options for feature_types at this stage:
0. SYM_SYM
\enumerate{
\item IDS_SYM
\item IDS_IDS
}
}
\details{
In the case of combination of homogeneous samples (SYM_SYM and IDS_IDS) we need
a "key" sample (IDS_SYM), in which case it'll try to infer symbols from ids or
viceversa. In case of absence of a key sample, this will throw an error, due
to incompatible features.
}
