% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{parse_cellsets}
\alias{parse_cellsets}
\title{Parse cellsets object to data.table}
\usage{
parse_cellsets(cellsets)
}
\arguments{
\item{cellsets}{list}
}
\value{
data.table of cellset keys, names and corresponding cell_ids
}
\description{
Gets the cellsets list and converts it to a tidy data.table
}
