% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{parse_rhapsody_matrix}
\alias{parse_rhapsody_matrix}
\title{Read BD Rhapsody expression_matrix.st files}
\usage{
parse_rhapsody_matrix(config, input_dir)
}
\arguments{
\item{input_dir}{A string, where the s3 object is going to be stored}
}
\value{
list containing \itemize{
  \item{"counts_list"}{named list of dgCMatrix per sample}
  \item{"annot"}{data.frame with gene symbols}
  }
}
\description{
Parses sparse count matrices generated by BD Rhapsody primary analysis.
}
