% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{perform_geosketch}
\alias{perform_geosketch}
\title{Perform geometric sketching}
\usage{
perform_geosketch(object, reduction, dims, num_cells)
}
\arguments{
\item{object}{Seurat object}

\item{reduction}{Reduced dimensions to use for sketching}

\item{dims}{Number of dimensions to use}

\item{num_cells}{Number of desired cells}
}
\value{
Seurat object downsampled to desired number of cells
}
\description{
See https://github.com/brianhie/geosketch
}
