% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-functions.R
\name{pipeline_error_handler}
\alias{pipeline_error_handler}
\title{Pipeline error handler}
\usage{
pipeline_error_handler(e)
}
\arguments{
\item{e}{error object}
}
\description{
Pretty prints errors, sends roport to the API, and uploads debug output to
S3 if the pipeline is running in a pod.
}
