% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-functions.R
\name{pipeline_heartbeat}
\alias{pipeline_heartbeat}
\title{Run heartbeat in the background}
\usage{
pipeline_heartbeat(task_token, aws_config)
}
\arguments{
\item{task_token}{character authorization token}

\item{aws_config}{list of parameters to access AWS step functions}
}
\description{
Sends a heartbeat to the state machine every 'wait_time' seconds Once the task
is completed the heartbeat will fail accordingly with a task timeout and exit
the loop and a new heartbeat will be set up by next task. This method is
invoked with \code{callr::r_bg} which creates a new process which does not inherit
the current workspace or memory, only the provided parameters; that's why we
need to re-import \code{tryCatchLog} & initialize states again.
}
