% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-6-prepare_experiment.R
\name{prepare_experiment}
\alias{prepare_experiment}
\title{Prepare experiment for upload to AWS}
\usage{
prepare_experiment(input, pipeline_config, prev_out)
}
\arguments{
\item{input}{The input object from the request}

\item{pipeline_config}{result of \code{load_config}}

\item{prev_out}{'output' slot from call to \code{create_seurat}}
}
\value{
prev_out \code{prev_out} with added slots 'scdata' containing merged
\code{SeuratObject} and 'qc_config' containing default config for QC steps
if it doesn't already exist.
}
\description{
\enumerate{
\item Adds metadata: cellsId, color_pool, and gene annotation
\item Prepares QC configuration
}
}
