% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata-seuratv4.R
\name{prepare_scdata_list_for_seurat_integration}
\alias{prepare_scdata_list_for_seurat_integration}
\title{prepare scdata list for seurat integration}
\usage{
prepare_scdata_list_for_seurat_integration(scdata_list, config, cells_id)
}
\arguments{
\item{scdata_list}{list of SeuratObjects}

\item{config}{list of configuration parameters}

\item{cells_id}{list of cells ids to keep}
}
\value{
scdata list
}
\description{
preprocess the scdata list before integration
}
