% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{read_10x_annotations}
\alias{read_10x_annotations}
\title{Load and parse annotations from the feature files}
\usage{
read_10x_annotations(annot_fpath, sample)
}
\arguments{
\item{annot_fpath}{chraracter path to features file}

\item{sample}{character sample id}
}
\value{
list of annotations data.frame
}
\description{
This function reads the features file into a data.frame, and takes steps
towards its standardization, converting it to a two-column data.frame, the first column
being the same as the rownames of the count matrix (ideally ensemblIDs), and
the second either the gene symbols if they are available or a copy of the gene IDs.
}
