% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{read_10x_h5_file}
\alias{read_10x_h5_file}
\title{Read h5 file}
\usage{
read_10x_h5_file(config, input_dir)
}
\arguments{
\item{config}{list of configuration parameters}

\item{input_dir}{character input dir}
}
\value{
list with counts and annotations
}
\description{
Calls read10x_h5
}
