% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{read_parse_annotations}
\alias{read_parse_annotations}
\title{Loads annotations and formats the annotation file with the column names required by format_annot}
\usage{
read_parse_annotations(annot_fpath, sample)
}
\arguments{
\item{annot_fpath}{Path to annotations file}

\item{sample}{Sample name}
}
\description{
Loads annotations and formats the annotation file with the column names required by format_annot
}
