% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{read_parse_files}
\alias{read_parse_files}
\title{Read Parse files. Calls readMTX}
\usage{
read_parse_files(config, input_dir)
}
\arguments{
\item{config}{experiment settings}

\item{input_dir}{}
}
\description{
Read Parse files. Calls readMTX
}
