% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{read_rhapsody_files}
\alias{read_rhapsody_files}
\title{Calls BD rhapsody data parsing functions}
\usage{
read_rhapsody_files(config, input_dir)
}
\arguments{
\item{input_dir}{A string, where the s3 object is going to be stored}
}
\value{
list containing \itemize{
  \item{"counts_list"}{named list of dgCMatrix per sample}
  \item{"annot"}{data.frame with gene symbols}
  }
}
\description{
Currently we only have support for sparse expression matrices.
}
