% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat-2-load_seurat.R
\name{reconstruct_seurat}
\alias{reconstruct_seurat}
\title{Reconstructs the Seurat object provided by the user}
\usage{
reconstruct_seurat(dataset_fpath)
}
\arguments{
\item{dataset_fpath}{The path to the r.rds file containing the Seurat object.}
}
\value{
reconstructed SeuratObject
}
\description{
The slots needed from the Seurat object are extracted and then their in-built types
are strictly checked before creating a new Seurat object. This reconstruction and type
checking prevents potentially malicious executable code  (e.g. type \code{closure} or \code{environment})
from being stored in the final Seurat object.
}
