% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{remove_filtered_cells}
\alias{remove_filtered_cells}
\title{For each sample, remove filtered cells from the Seurat object}
\usage{
remove_filtered_cells(scdata_list, cells_id)
}
\arguments{
\item{scdata_list}{list of SeuratObjects}

\item{cells_id}{list of cells ids to keep}
}
\value{
list of filtered SeuratObjects
}
\description{
For each sample, remove filtered cells from the Seurat object
}
