% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{remove_genes}
\alias{remove_genes}
\title{Remove genes from downstream analysis}
\usage{
remove_genes(scdata, exclude_groups, exclude_custom = list())
}
\arguments{
\item{scdata}{Seurat object}

\item{exclude_groups}{list of groups to exclude}

\item{exclude_custom}{list of custom (user provided) genes to exclude}
}
\value{
Seurat object without excluded genes
}
\description{
This function subsets the seurat object, removing genes to be excluded for
integration and downstream analysis.
}
\details{
It calls list_exclude_genes to build the list of genes to remove.
}
