% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-7-upload_to_aws.R
\name{remove_used_colors}
\alias{remove_used_colors}
\title{Remove used colors from pool}
\usage{
remove_used_colors(cellsets, color_pool)
}
\arguments{
\item{cellsets}{cell set object}

\item{color_pool}{list of colors}
}
\value{
color pool with used colors removed
}
\description{
Remove used colors from pool
}
