% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-7-embed_and_cluster.R
\name{replace_cell_class_through_api}
\alias{replace_cell_class_through_api}
\title{replace cell class by key}
\usage{
replace_cell_class_through_api(
  cell_class_object,
  api_url,
  experiment_id,
  cell_class_key,
  auth_JWT,
  ignore_ssl_cert
)
}
\arguments{
\item{cell_class_object}{list}

\item{api_url}{character}

\item{experiment_id}{character}

\item{cell_class_key}{character}

\item{auth_JWT}{character}

\item{ignore_ssl_cert}{boolean}
}
\description{
Replaces an entire cell class (group of cellsets) by key. Used to replace
clustering cellsets after re-clustering.
}
