% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-7-embed_and_cluster.R
\name{replace_cl_metadata_through_api}
\alias{replace_cl_metadata_through_api}
\title{Replace and Append Cell Metadata Through API}
\usage{
replace_cl_metadata_through_api(
  cl_metadata_cellsets,
  api_url,
  experiment_id,
  auth_JWT,
  ignore_ssl_cert
)
}
\arguments{
\item{cl_metadata_cellsets}{A list that will be appended to the existing cell sets.}

\item{api_url}{The base URL of the API.}

\item{experiment_id}{The unique identifier for the experiment to be modified.}

\item{auth_JWT}{The JSON Web Token used for authentication.}

\item{ignore_ssl_cert}{Boolean flag to indicate whether to ignore SSL certificate verification.}
}
\value{
The response from the API after performing the PATCH operation.
}
\description{
This function deletes cell-level metadata cellsets in an existing cellsets.json
file based on specified types. It then appends new cell-level metadata cellsets.
}
