% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-3-run_emptydrops.R
\name{run_emptydrops}
\alias{run_emptydrops}
\title{Run emptyDrops from DropletUtils}
\usage{
run_emptydrops(input, pipeline_config, prev_out)
}
\arguments{
\item{prev_out}{'output' slot from call to \code{load_cellranger}}
}
\value{
\code{prev_out} with added slot 'edrops' containing result of call to
\code{\link[DropletUtils]{emptyDrops}} for each sample.
}
\description{
Run emptyDrops from DropletUtils
}
