% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-functions.R
\name{run_gem2s_step}
\alias{run_gem2s_step}
\title{Run GEM2S step}
\usage{
run_gem2s_step(task_name, input, pipeline_config, prev_out)
}
\arguments{
\item{task_name}{character}

\item{input}{list
\itemize{
\item sampleIds character
\item sampleNames character
\item projectId character
}}

\item{pipeline_config}{list as defined by load_config}

\item{prev_out}{list output from previous step}
}
\value{
list of task results
}
\description{
Calls the corresponding task_name GEM2S step function.
}
\details{
The input list only contains experiment level parameters, such as project ID,
and sample names. It's only used for downloading user files.
}
