% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{run_lognorm_int_workflow}
\alias{run_lognorm_int_workflow}
\title{Find integration anchors after LogNormalize}
\usage{
run_lognorm_int_workflow(data.split, reduction, normalization, k.filter, npcs)
}
\arguments{
\item{data.split}{list of Seurat objects}

\item{reduction}{reduction method}

\item{normalization}{normalization method}

\item{k.filter}{number of neighbors (k) to use when filtering anchors}

\item{npcs}{number of PCs}
}
\value{
data.anchors to use for integration
}
\description{
This function finds the integration anchors for a list of Seurat objects
that has been normalized with the log-normalization method.
}
