% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{run_pca}
\alias{run_pca}
\title{Run PCA}
\usage{
run_pca(scdata)
}
\arguments{
\item{scdata}{Seurat object}
}
\value{
Seurat object with PCA slot
}
\description{
Performs FindVariableFeatures and ScaleData,
which are two required steps before RunPCA
}
