% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-functions.R
\name{run_qc_step}
\alias{run_qc_step}
\title{Run qc pipeline step}
\usage{
run_qc_step(
  scdata,
  config,
  tasks,
  task_name,
  cells_id,
  sample_id,
  ignore_ssl_cert,
  debug_config
)
}
\arguments{
\item{scdata}{Seurat Object}

\item{config}{list of configuration parameters}

\item{tasks}{list of pipeline tasks}

\item{task_name}{character}

\item{cells_id}{list of filtered cell ids}

\item{sample_id}{character}

\item{debug_config}{list of debug parameters}
}
\value{
list of task results
}
\description{
Calls the corresponding task_name QC pipeline step function.
}
