% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata-seuratv4.R
\name{run_seuratv4}
\alias{run_seuratv4}
\title{Run Seurat v4 integration}
\usage{
run_seuratv4(scdata_list, config, cells_id)
}
\arguments{
\item{scdata_list}{list of SeuratObjects}

\item{config}{list of configuration parameters}
}
\value{
normalized and integrated Seurat object
}
\description{
Integrates two or more samples using the Seurat v4 workflow.
It takes into account two different normaization methods: LogNormalize and SCTransform.
This function can also be used in combination with Geosketch. In this case,
the samples are first merged, then the whole dataset is downsampled using geometric sketching,
and the sketches are integrated. The integrated sketches are then used to learn
the integration transformation and apply it to the whole dataset.
}
