% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-helpers.R
\name{runClusters}
\alias{runClusters}
\title{Ruun clustering function}
\usage{
runClusters(clustering_method, resolution, data)
}
\arguments{
\item{clustering_method}{character}

\item{resolution}{numeric}

\item{data}{Seurat Object}
}
\value{
data.frame
}
\description{
Ruun clustering function
}
