% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_operations.R
\name{s3_copy_by_prefix}
\alias{s3_copy_by_prefix}
\title{Copy recursively all s3 objects matching prefix}
\usage{
s3_copy_by_prefix(
  source_bucket,
  source_prefix,
  destination_bucket,
  destination_prefix,
  aws_config,
  key_transform = identity
)
}
\arguments{
\item{source_bucket}{string}

\item{source_prefix}{string}

\item{destination_bucket}{string}

\item{destination_prefix}{string}

\item{aws_config}{list of AWS configuration}

\item{key_transform}{function to transform keys of objects before uploading, identity by default}
}
\description{
Copy recursively all s3 objects matching prefix
}
