% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{safe_cbind}
\alias{safe_cbind}
\title{Bind columns not failing if there's an empty data.table}
\usage{
safe_cbind(dt, ...)
}
\arguments{
\item{dt}{data.table}

\item{...}{columns to add}
}
\value{
data.table with new columns
}
\description{
Bind columns not failing if there's an empty data.table
}
