% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-4-score_doublets.R
\name{score_doublets}
\alias{score_doublets}
\title{Compute doublet score for count matrices}
\usage{
score_doublets(input, pipeline_config, prev_out)
}
\arguments{
\item{prev_out}{'output' slot from call to \code{run_emptydrops}}
}
\value{
\code{prev_out} with added slot 'doublet_scores' containing result of call to
\code{\link{compute_doublet_scores}} for each sample.
}
\description{
Compute doublet score for count matrices
}
