% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{sctransform_normalization}
\alias{sctransform_normalization}
\title{Perform normalization with SCTransform}
\usage{
sctransform_normalization(scdata)
}
\arguments{
\item{scdata}{Seurat object}
}
\value{
Seurat object normalized using SCTransform
}
\description{
This function exploit the Seurat::SCTransform function.
For further details see the \code{\link[Seurat:SCTransform]{Seurat::SCTransform()}}
documentation.
}
