% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata-seuratv4.R
\name{seuratv4_find_and_integrate_anchors}
\alias{seuratv4_find_and_integrate_anchors}
\title{Find and integrate anchors}
\usage{
seuratv4_find_and_integrate_anchors(
  scdata_list,
  cells_id,
  reduction,
  normalization,
  npcs,
  npcs_for_pca,
  nfeatures,
  scdata = NA,
  use_geosketch = FALSE
)
}
\arguments{
\item{scdata_list}{list of Seurat objects}

\item{cells_id}{list of cells ids to keep}

\item{reduction}{reduction method}

\item{normalization}{normalization method}

\item{npcs}{numer of principal components}

\item{nfeatures}{number of features}

\item{scdata}{merged Seurat object}

\item{use_geosketch}{boolean indicating if geosketch has to be run}
}
\value{
integrated Seurat object
}
\description{
This function find and integrate anchors according to the Seurat v4 workflow.
}
