% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata-seuratv4.R
\name{seuratv4_geosketch_find_and_integrate_anchors}
\alias{seuratv4_geosketch_find_and_integrate_anchors}
\title{Downsample and integrate sketches}
\usage{
seuratv4_geosketch_find_and_integrate_anchors(
  scdata_list,
  cells_id,
  reduction,
  perc_num_cells,
  normalization,
  npcs,
  npcs_for_pca,
  nfeatures,
  use_geosketch
)
}
\arguments{
\item{scdata_list}{list of Seurat objects}

\item{cells_id}{list of cells ids to keep}

\item{reduction}{reduction method}

\item{perc_num_cells}{percentage of cells to keep when using geosketch}

\item{normalization}{normalization method}

\item{npcs}{number of principal components}

\item{npcs_for_pca}{max number of PCs}

\item{nfeatures}{number of features}

\item{use_geosketch}{boolean indicating if geosketch has to be run}
}
\value{
integrated Seurat object
}
\description{
This function uses Geosketch to downsample the dataset, integrates the resulting
sketches using the Seurat v4 workflow, then learns
the integration transformation and applies it to the whole dataset.
}
