% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{shuffle_scdata}
\alias{shuffle_scdata}
\title{Shuffle the cells in a SeuratObject}
\usage{
shuffle_scdata(scdata)
}
\arguments{
\item{scdata}{SeuratObject}
}
\value{
SeuratObject
}
\description{
This function shuffles the cells in a Seurat object, to allow the UI to make
shuffled scatterplots, avoiding a sample bias when plotting.
}
