% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-7-embed_and_cluster.R
\name{sort_cluster_names}
\alias{sort_cluster_names}
\title{Sort cluster names}
\usage{
sort_cluster_names(strings)
}
\arguments{
\item{strings}{cluster names}
}
\value{
sorted vector
}
\description{
Sorts cluster names naturally, i.e. Cluster 1, Cluster 2, Cluster 10
}
