% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-functions.R
\name{start_heartbeat}
\alias{start_heartbeat}
\title{Start heartbeat as a background process}
\usage{
start_heartbeat(task_token, aws_config)
}
\arguments{
\item{task_token}{character authorization token}

\item{aws_config}{list of parameters to access AWS step functions}
}
\description{
messages inside the background process will ONLY be printed into
`/tmp/[out|err]`. To see them:
 1. log into the R container
 2. `cat /tmp/out` or `tail -f /tmp/out`
}
