% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-1-subset_seurat.R
\name{subset_experiment}
\alias{subset_experiment}
\title{Subset seurat object by the input cellset keys}
\usage{
subset_experiment(input, parent_data)
}
\arguments{
\item{input}{list of input parameters, containing cellSetKeys to subset}

\item{parent}{list containing parent scdata and parsed cellsets}
}
\value{
subset seurat object
}
\description{
This function takes the cellset keys sent by the API, extracts the cell_ids
that belong to them, subsets the seurat object and removes all unnecessary
data from it.
}
