% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-helpers.R
\name{subset_ids}
\alias{subset_ids}
\title{Subset Seurat object given cell_ids}
\usage{
subset_ids(scdata, cells_id)
}
\arguments{
\item{scdata}{Seurat rds object downloaded from S3}

\item{cells_id}{numeric vector of cell ids to extract}
}
\value{
subsetted Seurat object
}
\description{
Cell Ids are an internal representation used by Cellenics. They are stored in
the \code{meta.data} slot of the Seurat object. This function subsets a Seurat
object given a vector of cell ids.
}
