% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-helpers.R
\name{subset_safe}
\alias{subset_safe}
\title{Safely subset Seurat object}
\usage{
subset_safe(scdata, cells)
}
\arguments{
\item{scdata}{Seurat object to filter}

\item{cells}{cell barcodes to subset the object with}
}
\value{
subsetted Seurat Object
}
\description{
Subset safe allows us to attempt to subset a seurat object even with an empty
list. This function exists for the case when the user over-filters the object,
and we need to return something that'd allow the user to realize that they are
filtering all the cells, while maintaining certain seurat functionality.
}
\details{
It's a questionable function and it should be questioned.
}
