% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{sym_to_ids}
\alias{sym_to_ids}
\title{Convert symbols to ids using key data.frame}
\usage{
sym_to_ids(sample_annot, annot_with_ids)
}
\arguments{
\item{sample_annot}{data.frame of annotations}

\item{annot_with_ids}{data.frame of annotations with IDs. Key data.frame}
}
\value{
data.frame of annotations
}
\description{
This function tries to convert symbols to ids using the key data.frame created
with \code{make_annot_with_ids}. In case there's no ID available, it keeps the
existing symbol.
}
