% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-1-copy_s3_objects.R
\name{translate_cell_sets}
\alias{translate_cell_sets}
\title{Replaces sample cell sets keys with their matched values in sample_ids_map}
\usage{
translate_cell_sets(cell_sets, sample_ids_map)
}
\arguments{
\item{cell_sets}{original cell sets with old sample ids as keys}

\item{sample_ids_map}{data.table of parent/subset sample id map}
}
\value{
cell sets with new sample ids as keys instead of old ones
}
\description{
Replaces sample cell sets keys with their matched values in sample_ids_map
}
