% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy-1-copy_s3_objects.R
\name{translate_sample_ids}
\alias{translate_sample_ids}
\title{Replaces sample ids with their matched values in sample_ids_map}
\usage{
translate_sample_ids(old_sample_ids, sample_ids_map)
}
\arguments{
\item{old_sample_ids}{old sample ids vector}

\item{sample_ids_map}{data.table of parent/subset sample id map}
}
\value{
vector with new sample ids matching the order of old_sample_ids
}
\description{
Replaces sample ids with their matched values in sample_ids_map
}
