% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_data.R
\name{unflatten_cell_sets}
\alias{unflatten_cell_sets}
\title{Load cellsets object flattened (as is done by load_cellsets)
and change it back to the original shape
This is useful to prepare the cell sets to upload back into s3}
\usage{
unflatten_cell_sets(cell_sets)
}
\arguments{
\item{s3}{paws::s3 object}

\item{pipeline_config}{list}

\item{experiment_id}{character}
}
\value{
cellsets list
}
\description{
Load cellsets object flattened (as is done by load_cellsets)
and change it back to the original shape
This is useful to prepare the cell sets to upload back into s3
}
