% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-7-upload_to_aws.R
\name{upload_to_aws}
\alias{upload_to_aws}
\title{Upload Seurat and cellsets objects to aws}
\usage{
upload_to_aws(input, pipeline_config, prev_out)
}
\arguments{
\item{input}{The input object from the request}

\item{pipeline_config}{result of \code{load_config}}

\item{prev_out}{list with results appended in each gem2s task}
}
\value{
list with experiment parameters
}
\description{
Upload Seurat and cellsets objects to aws
}
