% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withErrorTracing.R
\name{withErrorTracing}
\alias{withErrorTracing}
\title{Add Error Tracing to tryCatch}
\usage{
withErrorTracing(expr, silentSuccess = FALSE, skip_pattern = "^multi_assign")
}
\arguments{
\item{expr}{expression to evaluate}

\item{skip_pattern}{grep pattern. Will not print calls including or after
first match of \code{skip_pattern}}
}
\description{
Add Error Tracing to tryCatch
}
