% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-functions.R
\name{wrapper}
\alias{wrapper}
\title{calls the appropriate process, QC or gem2s}
\usage{
wrapper(input, pipeline_config)
}
\arguments{
\item{input}{list with parsed input json. should contain:
\itemize{
\item task name
\item current sample_id
\item extra config parameters
}}

\item{pipeline_config}{list as generated by load_config}
}
\value{
character message id
}
\description{
calls the appropriate process, QC or gem2s
}
