% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-1-subset_seurat.R
\name{add_new_sample_ids}
\alias{add_new_sample_ids}
\title{Add new sample ids to the subset Seurat Object}
\usage{
add_new_sample_ids(subset_scdata, sample_id_map)
}
\arguments{
\item{sample_id_map}{data.table of parent/subset sample id map}

\item{scdata}{Seurat Object}
}
\value{
SeuratObject with new sample ids
}
\description{
Add new sample ids to the subset Seurat Object
}
