% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-7-upload_to_aws.R
\name{build_scratchpad_cellsets}
\alias{build_scratchpad_cellsets}
\title{Create cellsets using scratchpad information from parent cellsets}
\usage{
build_scratchpad_cellsets(color_pool, subset_cellsets)
}
\arguments{
\item{color_pool}{list of colors to use}

\item{subset_cellsets}{cell set resulting from parent cell set filtering}
}
\value{
cell set filled with scratchpad information from the parent cell set
}
\description{
This function creates cellsets with scratchpad information including only
the cell ids in the subset experiment. If all cell ids from a custom cellset
of the parent experiments are filtered out in the subset experiment,
then that custom cellset is not included in the subset cell set.
}
