% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-helpers.R
\name{calc_filter_stats}
\alias{calc_filter_stats}
\title{Calculates statistics before/after filter step}
\usage{
calc_filter_stats(seurat_obj, tmp_sample)
}
\arguments{
\item{seurat_obj}{\code{SeuratObject}}

\item{tmp_sample}{sample name in \code{seurat_obj$samples} to compute statistics for}
}
\value{
list with \itemize{
\item{"num_cells"}{Number of cells in sample}
\item{"total_genes"}{Number of detected genes in sample}
\item{"median_genes"}{Median number of genes detected per cell}
\item{"median_umis"}{Median number of counts per cell}
}
}
\description{
Calculates statistics before/after filter step
}
