% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-4-score_doublets.R
\name{compute_sample_doublet_scores}
\alias{compute_sample_doublet_scores}
\title{Compute doublets scores per sample.}
\usage{
compute_sample_doublet_scores(sample_counts, technology, parse_kit = "WT")
}
\arguments{
\item{sample_counts}{Sparse matrix with the counts for one sample.}

\item{technology}{Technology used to generate the data (e.g. 10x, Parse).}

\item{parse_kit}{Kit used to generate the data (specific to Parse data).}
}
\value{
data.frame with doublet scores and assigned classes
}
\description{
Compute doublets scores per sample.
}
