% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-1-subset_seurat.R
\name{diet_scdata}
\alias{diet_scdata}
\title{Remove all unnecessary data from the parent seurat object}
\usage{
diet_scdata(scdata)
}
\arguments{
\item{scdata}{SeuratObject}
}
\value{
leaner SeuratObject
}
\description{
Seurat::DietSeurat is not able to remove certain slots from a seurat object.
This function also removes elements from the misc slot which are not necessary
}
