% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-7-upload_to_aws.R
\name{filter_parent_cellsets}
\alias{filter_parent_cellsets}
\title{Filter parent cellsets}
\usage{
filter_parent_cellsets(parent_cellsets, cell_ids_to_keep)
}
\arguments{
\item{parent_cellsets}{data.tabble with cellsets from the parent experiment}

\item{cell_ids_to_keep}{integer vector of cell ids to keep}
}
\value{
filtered cellssets data.table
}
\description{
This function filters the parent cellsets removing clusters, which will be
recalculated in the subset experiment. It also filters cell ids not present
after the subset.
}
