% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{get_feature_types}
\alias{get_feature_types}
\title{Determine the type of features in the annot data frame}
\usage{
get_feature_types(annot)
}
\arguments{
\item{annot}{data.frame read from features file}
}
\value{
character vector indicating feature types
}
\description{
Classifies the features file columns into either ensemblIds or symbols
}
