% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_user_files.R
\name{get_feature_types}
\alias{get_feature_types}
\title{extract_feature_types}
\usage{
get_feature_types(annot)
}
\arguments{
\item{annot}{}
}
\value{
-1 is SYMBOl/IDS
0 is SYMBOL/ SYMBOL
1 is IDS/SYMBOL
2 is IDS/IDS
}
\description{
Determines the type of an annot data frame.
Classifies the columns into either ensemblIds or symbols, and extracts a number that represents the combination.
}
