% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{learn_from_sketches}
\alias{learn_from_sketches}
\title{Learn integration transformation from sketches}
\usage{
learn_from_sketches(
  scdata,
  scdata_sketch,
  scdata_sketch_integrated,
  method,
  npcs
)
}
\arguments{
\item{scdata}{Seurat object}

\item{scdata_sketch}{Sketched Seurat object}

\item{scdata_sketch_integrated}{Sketched integrated Seurat object}

\item{method}{Reduction method}

\item{npcs}{Number of PCs}
}
\value{
Integrated Seurat object with original number of cells
}
\description{
Uses the integrated sketches to learn the integration transformation and
apply it to the whole dataset
}
