% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gem2s-2-load_cellranger_files.R
\name{load_cellranger_files}
\alias{load_cellranger_files}
\title{Read input folder of 10x data}
\usage{
load_cellranger_files(input, pipeline_config, prev_out)
}
\arguments{
\item{input}{The input object from the request}

\item{pipeline_config}{result of \code{load_config}}

\item{prev_out}{list with experiment configuration settings}
}
\value{
list with 'output' slot containing \itemize{
\item{"counts_list"}{named list of dgCMatrix per sample}
\item{"annot"}{data.frame from reading features.tsv.gz for each sample}
}
}
\description{
Read input folder of 10x data
}
