% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc-6-integrate_scdata.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize data according to the specific normalization method}
\usage{
normalize_data(scdata, normalization_method, integration_method, nfeatures)
}
\arguments{
\item{scdata}{SeuratObject}

\item{normalization_method}{normalization method}

\item{integration_method}{integration method}

\item{nfeatures}{number of features to pass to Seurat::FindVariableFeatures()}
}
\value{
normalized and scaled SeuratObject
}
\description{
This function normalize the data taking into account the integration method.
}
